Index

StartUpDG.MeshData

StartUpDG.MeshPlotter

StartUpDG.RefElemData

StartUpDG.RefElemData

StartUpDG.RefElemData

StartUpDG.RefElemData

StartUpDG.VertexMeshPlotter

StartUpDG.boundary_face_centroids

StartUpDG.ck45

StartUpDG.connect_mesh

StartUpDG.estimate_h

StartUpDG.find_face_nodes

StartUpDG.hybridized_SBP_operators

StartUpDG.inverse_trace_constant

StartUpDG.make_periodic

StartUpDG.readGmsh2D

StartUpDG.tag_boundary_faces

StartUpDG.uniform_mesh

Functions

StartUpDG.MeshData — Type

struct MeshData{Dim, Tv, Ti}

MeshData: contains info for a high order piecewise polynomial discretization on an unstructured mesh.

Example:

N, K1D = 3, 2
rd = RefElemData(Tri(), N)
VXY, EToV = uniform_mesh(Tri(), K1D)
md = MeshElemData(VXY, EToV, rd)
@unpack x, y = md

StartUpDG.MeshPlotter — Type

MeshPlotter(rd::RefElemData, md::RefElemData)

Plot recipe to plot a (possibly curved) quadrilateral or triangular mesh. Usage: plot(MeshPlotter(...))

StartUpDG.RefElemData — Type

struct RefElemData{Dim, ElemShape <: AbstractElemShape, ApproximationType, Nfaces, Tv, VQ, VF, MM, P, D, L}

RefElemData: contains info (interpolation points, volume/face quadrature, operators) for a high order nodal basis on a given reference element.

Example:

N = 3
rd = RefElemData(Tri(), N)
@unpack r, s = rd

StartUpDG.RefElemData — Method

function RefElemData(elem; N, kwargs...)
function RefElemData(elem, approxType; N, kwargs...)

Keyword argument constructor for RefElemData (to "label" N via rd = RefElemData(Line(), N=3))

StartUpDG.RefElemData — Method

RefElemData(elem::Line, N;
            quad_rule_vol = quad_nodes(elem, N+1))
RefElemData(elem::Union{Tri, Quad}, N;
             quad_rule_vol = quad_nodes(elem, N),
             quad_rule_face = gauss_quad(0, 0, N))
RefElemData(elem::Union{Hex, Tet}, N;
             quad_rule_vol = quad_nodes(elem, N),
             quad_rule_face = quad_nodes(Quad(), N))
RefElemData(elem; N, kwargs...) # version with keyword args

Constructor for RefElemData for different element types.

StartUpDG.RefElemData — Method

function RefElemData(elementType::Line, approxType::SBP, N)
function RefElemData(elementType::Quad, approxType::SBP, N)
function RefElemData(elementType::Hex,  approxType::SBP, N)
function RefElemData(elementType::Tri,  approxType::SBP, N)

SBP reference element data for Quad(), Hex(), and Tri() elements.

For Line(), Quad(), and Hex(), approxType is SBP{TensorProductLobatto}.

For Tri(), approxType can be SBP{Kubatko{LobattoFaceNodes}}, SBP{Kubatko{LegendreFaceNodes}}, or SBP{Hicken}.

StartUpDG.VertexMeshPlotter — Type

VertexMeshPlotter((VX, VY), EToV, fv)
VertexMeshPlotter(triout::TriangulateIO)

Plot recipe to plot a quadrilateral or triangular mesh. Usage: plot(VertexMeshPlotter(...))

StartUpDG.boundary_face_centroids — Method

function boundary_face_centroids(md)

Returns face centroids and boundary_face_ids on the boundaries of the domain given by md::MeshData.

StartUpDG.ck45 — Method

ck45()

Returns coefficients rka,rkb,rkc for the 4th order 5-stage low storage Carpenter/Kennedy Runge Kutta method. Coefficients evolve the residual, solution, and local time, e.g.,

Example

res = rk4a[i]*res + dt*rhs # i = RK stage
@. u += rk4b[i]*res

StartUpDG.connect_mesh — Method

connect_mesh(EToV,fv)

Initialize element connectivity matrices, element to element and element to face connectivity.

Inputs:

EToV is a K by Nv matrix whose rows identify the Nv vertices

which make up an element.

fv (an array of arrays containing unordered indices of face vertices).

Output: FToF, length(fv) by K index array containing face-to-face connectivity.

StartUpDG.estimate_h — Method

estimate_h(rd::RefElemData, md::MeshData)
estimate_h(e, rd::RefElemData, md::MeshData) # e = element index

Estimates the mesh size via min sizeofdomain * |J|/|sJ|, since |J| = O(hᵈ) and |sJ| = O(hᵈ⁻¹).

StartUpDG.find_face_nodes — Function

function find_face_nodes(elem, r, s, tol=50*eps())
function find_face_nodes(elem, r, s, t, tol=50*eps())

Given volume nodes r, s, t, finds face nodes. Note that this function implicitly defines an ordering on the faces.

StartUpDG.hybridized_SBP_operators — Method

function hybridized_SBP_operators(rd::RefElemData{DIMS})

Constructs hybridized SBP operators given a RefElemData. Returns operators Qrsth..., VhP, Ph.

StartUpDG.inverse_trace_constant — Method

function inverse_trace_constant(rd::RefElemData)

Returns the degree-dependent constant in the inverse trace equality over the reference element (as reported in "GPU-accelerated dG methods on hybrid meshes" by Chan, Wang, Modave, Remacle, Warburton 2016).

Can be used to estimate dependence of maximum stable timestep on degree of approximation.

StartUpDG.make_periodic — Method

make_periodic(md::MeshData{Dim}, is_periodic...) where {Dim}
make_periodic(md::MeshData{Dim}, is_periodic = ntuple(x->true,Dim)) where {Dim}
make_periodic(md::MeshData, is_periodic = true)

Returns new MeshData such that the node maps mapP and face maps FToF are now periodic. Here, is_periodic is a tuple of Bool indicating whether or not to impose periodic BCs in the x,y, or z coordinate.

StartUpDG.readGmsh2D — Method

function readGmsh2D(filename)

reads triangular GMSH 2D file format 2.2 0 8. returns VX,VY,EToV

Examples

VXY, EToV = readGmsh2D("eulerSquareCylinder2D.msh")

StartUpDG.tag_boundary_faces — Method

function tag_boundary_faces(md, boundary_name::Symbol = :entire_boundary)
function tag_boundary_faces(md, boundary_list::Dict{Symbol, <:Function})

When called without arguments, just returns Dict(:entire_boundary => boundary_faces)`.

Example usage:

julia> rd = RefElemData(Tri(), N=1)
julia> md = MeshData(uniform_mesh(Tri(), 2)..., rd)
julia> on_bottom_boundary(x, y, tol = 1e-13) = abs(y+1) < tol
julia> on_top_boundary(x, y, tol = 1e-13) = abs(y-1) < tol
julia> tag_boundary_faces(Dict(:bottom => on_bottom_boundary,
                               :top    => on_top_boundary), md)

StartUpDG.uniform_mesh — Method

uniform_mesh(elem::Line,Kx)
uniform_mesh(elem::Tri,Kx,Ky)
uniform_mesh(elem::Quad,Kx,Ky)
uniform_mesh(elem::Hex,Kx,Ky,Kz)  
uniform_mesh(elem, K)

Uniform Kx (by Ky by Kz) mesh on $[-1,1]^d$, where d is the spatial dimension. Returns (VX,VY,VZ), EToV. When only one K is specified, it assumes a uniform mesh with K elements in each coordinate direction.

K can also be specified using a keyword argument K1D, e.g., uniform_mesh(elem; K1D = 16).